Converts a list of time bin rules to a table of rules for each day of the week

to7daybins created a table of 7 rows, with rules per day.

to7daybins(rules)

Arguments

rules

A list of lists of rules, each sublist must contain 4 variables, start and end as the start and end times (24h format) of the time bin, days as a vector of days to apply this time bin to (1 for Sunday, ..., 7 for Saturday), and tag as the name of the time bin.

Value

Returns a table of time bin rules for each day of the week.

Examples

if (FALSE) {

rules = list(
    list(start='6:30',  end= '9:00',  days = 1:5, tag='MR'),
    list(start='15:00', end= '18:00', days = 2:5, tag='ER'))

to7daybins(rules)
}